Final Sub
With being sick and other finals, I ended up just cutting the wifi to my desktop so I could work uninterrupted. I realize that makes it look like I've only just now done all this work, but this has been quite the journey. Some things still aren't working but it's 2 AM right now and it's gonna be working as much as it can at this point. Thank you so much for a fun semester Professor! I'll likely keep working on this over the summer, get it finalized. Have a great summer!